+++ Frontend/src/context/AuthProvider.tsx
@@ -1,16 +1,6 @@
-import React, { createContext, useContext, useState, useEffect } from 'react';
+import React, { useState, useEffect } from 'react';
+import { AuthContext } from './AuthContext';
 import api from '../utils/api';
-
-interface AuthContextType {
-  isAuthenticated: boolean;
-  accessToken: string | null;
-  refreshToken: string | null;
-  login: (accessToken: string, refreshToken: string) => void;
-  logout: () => void;
-  api: typeof api;
-}
-
-export const AuthContext = createContext<AuthContextType | null>(null);
 
 export const AuthProvider: React.FC<{children: React.ReactNode}> = ({ children }) => {
   const [accessToken, setAccessToken] = useState<string | null>(localStorage.getItem('auth_token'));
@@ -56,12 +46,4 @@ export const AuthProvider: React.FC<{children: React.ReactNode}> = ({ children }
       {children}
     </AuthContext.Provider>
   );
-};
-
-export const useAuth = () => {
-  const context = useContext(AuthContext);
-  if (!context) {
-    throw new Error('useAuth must be used within an AuthProvider');
-  }
-  return context;
 };
