+++ Frontend/src/context/__tests__/AuthContext.test.tsx
@@ -1,6 +1,7 @@
 import { describe, it, expect, vi, beforeEach, afterEach } from 'vitest'
 import { render, screen, act } from '@testing-library/react'
-import { AuthProvider, useAuth } from '../AuthContext'
+import { AuthProvider } from '../AuthProvider'
+import { useAuth } from '../../hooks/useAuth'
 
 // Mock the api module
 vi.mock('../../utils/api', () => ({
