+++ Frontend/src/pages/clubs/CreateClub.tsx
@@ -1,6 +1,6 @@
 import { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
-import { useAuth } from '../../context/AuthContext';
+import { useAuth } from '../../hooks/useAuth';
 import Layout from '../../components/layout/Layout';
 
 const CreateClub = () => {
