+++ Frontend/src/utils/api.ts
@@ -8,9 +8,9 @@ const api = axios.create({
 });
 
 let isRefreshing = false;
-let failedQueue: { resolve: (token: string) => void; reject: (error: any) => void; }[] = [];
+let failedQueue: { resolve: (token: string) => void; reject: (error: unknown) => void; }[] = [];
 
-const processQueue = (error: any | null, token: string | null = null) => {
+const processQueue = (error: unknown, token: string | null = null) => {
   failedQueue.forEach(prom => {
     if (error) {
       prom.reject(error);
